# Faster string reversal when source/dest buffers are distinct