`pp_reverse()` has traditionally operated on a single string buffer,
reversing it in place.

When `pp_reverse()` takes a single SV argument, whose buffer cannot be
swiped by `sv_setsv_flags()`, a straight copy of the buffer is taken
and then that is reversed. In such cases, two complete passes over the
entire string are needed to reverse a non-UTF8 string, and reversing a
UTF8 string takes three complete passes.

This PR enables `pp_reverse()` to apply the "swipe" test itself
and, for straightforward copy cases, to avoid calling `sv_setsv_flags()`.
Instead, it does the work of preparing the `TARG` SV and reverse copies
the string (UTF8 or otherwise) in a single pass.

Non-UTF8 performance is more than doubled. UTF8 performance is improved
by ~40%. (This is using `gcc version 12.2.0` on Linux running on a midrange Zen3 CPU.)

* `my $x = "X"x(1024*1000*10); my $y; for (0..1_000) { $y = reverse $x }`

PATCHED
```
          2,300.42 msec task-clock                       #    0.994 CPUs utilized          
                 2      context-switches                 #    0.869 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,472      page-faults                      #    1.075 K/sec                  
    10,523,888,668      cycles                           #    4.575 GHz                    
         3,407,581      stalled-cycles-frontend          #    0.03% frontend cycles idle   
        17,098,625      stalled-cycles-backend           #    0.16% backend cycles idle    
    61,520,603,431      instructions                     #    5.85  insn per cycle         
    10,255,033,319      branches                         #    4.458 G/sec                  
```       

BLEAD       
```
          5,693.74 msec task-clock                       #    0.996 CPUs utilized          
                 4      context-switches                 #    0.703 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,471      page-faults                      #  433.985 /sec                   
    26,417,106,165      cycles                           #    4.640 GHz                    
         4,123,476      stalled-cycles-frontend          #    0.02% frontend cycles idle   
        16,843,163      stalled-cycles-backend           #    0.06% backend cycles idle    
    41,985,580,059      instructions                     #    1.59  insn per cycle         
     5,210,998,998      branches                         #  915.215 M/sec                  
```
       
* `my $x = "\x{263A}\x{263A}x\x{263A}y\x{263A}"x(1024*1000); my $y; for (0..1_000) { $y = reverse $x }`

PATCHED
```
         13,473.52 msec task-clock                       #    0.999 CPUs utilized          
                 8      context-switches                 #    0.594 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,869      page-faults                      #  212.936 /sec                   
    64,319,489,828      cycles                           #    4.774 GHz                    
        19,171,494      stalled-cycles-frontend          #    0.03% frontend cycles idle   
        29,132,509      stalled-cycles-backend           #    0.05% backend cycles idle    
   139,440,302,209      instructions                     #    2.17  insn per cycle         
    32,809,474,409      branches                         #    2.435 G/sec                  
```

BLEAD
```
         21,833.47 msec task-clock                       #    0.999 CPUs utilized          
                18      context-switches                 #    0.824 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,362      page-faults                      #  108.183 /sec                   
    99,685,066,602      cycles                           #    4.566 GHz                    
        17,824,958      stalled-cycles-frontend          #    0.02% frontend cycles idle   
        34,272,539      stalled-cycles-backend           #    0.03% backend cycles idle    
   378,833,407,708      instructions                     #    3.80  insn per cycle         
    72,845,191,845      branches                         #    3.336 G/sec                  
```
---------------------------------------------------------------------------------
* This set of changes requires a perldelta entry, and it will be included shortly.