@richardleach 2025-02-18
Note: I have a separate branch tinkering with (1) minor variations on the current algorithms (2) the portable code mentioned in https://github.com/Perl/perl5/issues/20692. Likely there will be a subsequent PR for at least the byte-reversal algorithm.

I don't know how to significantly improve the UTF8 algorithms - any suggestions welcome!