@richardleach 2025-02-23
I added a branch for the single byte UFT8 case while squashing bugs. `gcc` timings on that benchmark :
- `my $x = "\x{263A}\x{263A}x\x{263A}y\x{263A}"x(1024*1000); my $y; for (0..1_000) { $y = reverse $x }`
Blead was `99,685,066,602 cycles`.
PR      was `64,319,489,828 cycles`.
PR now `52,251,356,819 cycles`:
```
         11,450.39 msec task-clock                       #    0.997 CPUs utilized          
                 8      context-switches                 #    0.699 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,873      page-faults                      #  250.908 /sec                   
    52,251,356,819      cycles                           #    4.563 GHz                    
         8,948,516      stalled-cycles-frontend          #    0.02% frontend cycles idle   
        26,727,744      stalled-cycles-backend           #    0.05% backend cycles idle    
   145,588,067,775      instructions                     #    2.79  insn per cycle         
    32,808,946,621      branches                         #    2.865 G/sec
```
Compiling with _clang-11_, the "before" times are both about the same as with `gcc`, as mentioned in the original PR message. However, _clang_ does a lot better with the new copy algorithms. It manages to vectorizes where _gcc_ (including later versions up to _trunk_) does not manage to do so.
- `my $x = "X"x(1024*1000*10); my $y; for (0..1_000) { $y = reverse $x }`
Blead used about `26,417,106,165 cycles`.
gcc  used about `  10,523,888,668 cycles`.
clang uses about ` 3,073,025,623 cycles`.

```
            699.47 msec task-clock                       #    0.924 CPUs utilized          
                 0      context-switches                 #    0.000 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,456      page-faults                      #    3.511 K/sec                  
     3,073,025,623      cycles                           #    4.393 GHz                    
         1,397,171      stalled-cycles-frontend          #    0.05% frontend cycles idle   
        15,385,164      stalled-cycles-backend           #    0.50% backend cycles idle    
     8,986,055,265      instructions                     #    2.92  insn per cycle         
       324,660,921      branches                         #  464.155 M/sec  
```
- `my $x = "\x{263A}\x{263A}x\x{263A}y\x{263A}"x(1024*1000); my $y; for (0..1_000) { $y = reverse $x }`
Blead used about `99,685,066,602 cycles`.
gcc  used about `  52,251,356,819 cycles`.
clang uses about `33,720,311,746 cycles`.
```
          7,474.72 msec task-clock                       #    0.997 CPUs utilized          
                13      context-switches                 #    1.739 /sec                   
                 0      cpu-migrations                   #    0.000 /sec                   
             2,346      page-faults                      #  313.858 /sec                   
    33,720,311,746      cycles                           #    4.511 GHz                    
        11,169,631      stalled-cycles-frontend          #    0.03% frontend cycles idle   
        25,214,673      stalled-cycles-backend           #    0.07% backend cycles idle    
   182,481,212,299      instructions                     #    5.41  insn per cycle         
    36,907,239,014      branches                         #    4.938 G/sec
```