@thesamesam 2025-02-23
> It manages to vectorizes where gcc (including later versions up to trunk) does not manage to do so.

Please do consider filing a bug with preprocessed sources to GCC for that (even better with a simple testcase). Thanks.