@richardleach 2025-03-02
> > It manages to vectorizes where gcc (including later versions up to trunk) does not manage to do so.
> 
> Please do consider filing a bug with preprocessed sources to GCC for that (even better with a simple testcase; just a function or function(s) which clearly don't vectorise well compared to Clang in godbolt is fine. no need for it to be benchmarkable). Thanks.

https://godbolt.org/z/jz6fxKKhh probably captures this. A simple `reverse` loop: _gcc_ outputs simple code, but _clang_ goes to town.

```
#include <stddef.h>
void
reverse(char * outp, const char * inps, const char * inpe) {
    while (inpe != inps)
        *outp++ = *--inpe;
}
```
I got the impression that alias checking under gcc is a known problem and is covered by a metabug, but I can't find the link to it right now. I didn't fancy wading though all the linked issues to check for duplicates at this time. Feel free to report this case though if you believe it to be useful to the gcc folks.

I've got a follow-up commit planned that would give gcc and clang similar performance for string reversal in Perl.