@tonycoz 2025-03-02
`-O3` and a semi-recent `-march` (`skylake`, `x86-64-v2`) lets gcc vectorize this.