@richardleach 2025-03-03
> `-O3` and a semi-recent `-march` (`skylake`, `x86-64-v2`) lets gcc vectorize this.

I gather that building for `x86-64-v2` is gaining traction, so that's pretty cool. Are any major distros building with gcc at `-O3` though?