@tonycoz 2025-03-03
> If those buffers are distinct, the `restrict` keyword is the way to go.

`restrict` on `outp` for the godbolt example doesn't enable the vectorization, it just eliminates the fallback to the direct implementation on overlap.