@xenu 2025-03-03
It does in GCC trunk, and indeed, according to WIP [GCC 15 release notes](https://gcc.gnu.org/gcc-15/changes.html), it is supposed to vectorize more aggressively at -O2.