@richardleach 2025-03-03
> If those buffers are distinct, the `restrict` keyword is the way to go.

I had that in originally, but it seemed to break g++ builds?